Change link behavior to always link through.
Now the menu is only shown on hover. This change makes the dropdown more compatible with a mobile first approach where the menu is actually a shortcut on mobile, but can be a menu on desktop. (This requires that these menu items be shown somewhere more suitable on mobile though).